GUI: Ensure that event-tick-render phases don't interleave each other
Previously, the GUI was updated in the same time as drawn. Some draw
functions were order-dependent because of that.

This caused a bug where the planet placeholder could be moved even when
an existing planet was focused, because it was drawn before
SimulationView was updated.

This patch makes that updates are done always before rendering, which
fixes this behavior.