Gfx: Add ClipViewScope class
This is a RAII class which sets up a view which transform the
coordinates from (position, position + size) to (0, size) and doesn't
render anything outside the clip rect. This is currently used in Widget,
which will be ported to ClipViewScope in the next commit.

You can set how the the view will be treated according to the parent
(e.g current when creating ClipViewScope) view (overwrite or
intersect). This can be used to limit Widget's rendering area to the
parent Container.