Object: Move storing old most attracting object to before_update
When it was in clear_forces(), the wrong most attracting object was
stored because clear_forces() is called twice. Now, in
recalculate_trails_with_offset() we see the actual old MAO
(i.e from the previous tick).

This makes first vertex of trail recalculated properly because of the
offset applied in the first tick.